Rename ballot manifest upload saga.
